+++ test/loops_test.rb
@@ -13,6 +13,34 @@ class LoopsTest < Minitest::Test
     OUTPUT
   end
 
+  def test_while_false
+    assert_change <<~INPUT, ''
+      while false
+        puts 'Hello, world!'
+      end
+    INPUT
+  end
+
+  def test_until_false
+    assert_change <<~INPUT, <<~OUTPUT
+      until false
+        puts 'Hello, world!'
+      end
+    INPUT
+      loop do
+      puts "Hello, world!"
+      end
+    OUTPUT
+  end
+
+  def test_until_true
+    assert_change <<~INPUT, ''
+      until true
+        puts 'Hello, world!'
+      end
+    INPUT
+  end
+
   def test_for
     assert_change <<~INPUT, <<~OUTPUT
       for foo in [1, 2, 3]